Move httputils.PageQuery to api.PageQuery
