+++ lib/node/runner/api/pagequery.go
@@ -1,4 +1,4 @@
-package httputils
+package api
 
 import (
 	"fmt"
